Вариант с ошибкой в отправке
