+++ js/server.js
@@ -1,28 +1,6 @@
 const URL_SERVER = 'https://27.javascript.pages.academy/kekstagram-simple' ;
 const URL_PHOTOS = `${URL_SERVER}/data`;
-const TIME = 5000;
-const showAlert = (message) => {
-  const alertContainer = document.createElement('div');
-  alertContainer.style = `z-index:100;
-    position: absolute;
-    width: 100%;
-    left: 50%;
-    top: 50%;
-    padding: 10px 3px;
-    font-size: 18px;
-    line-height: 20px;
-    text-align: center;
-    background-color: red;
-    transform: translate(-50%, -50%);`;
-
-  alertContainer.textContent = message;
-
-  document.body.append(alertContainer);
-
-  setTimeout(() => {
-    alertContainer.remove();
-  }, TIME);
-};
+import { showAlert } from './show-alert.js';
 
 const getData = (onSuccess) => {
   fetch(
