+++ js/util.js
@@ -45,6 +45,8 @@ function getListElements(counter) {
   return newArray;
 }
 
+const isEscapeKey = (evt) => evt.key === 'Escape';
+
 // const showAlert = (message) => {
 //   const alertContainer = document.createElement('div');
 //   alertContainer.style.zIndex = '100';
@@ -71,4 +73,5 @@ export {getRandomArrayElement};
 export {getRandomIntInclusive};
 export {checkMaxString};
 export {getListElements};
+export {isEscapeKey};
 // export {showAlert};
