+++ js/validity.js
@@ -45,6 +45,7 @@ import { closePopupEscKeydown } from './form-img.js';
 import { sendData } from './server.js';
 import { isEscapeKey } from './form-img.js';
 
+
 const formElement = document.getElementById('upload-select-image');
 const btnSubmitElement = document.querySelector('.img-upload__submit');
 const messageSuccessTemplateElement = document.querySelector('#success').content.querySelector('.success');
